+++ website/src/components/Logo.tsx
@@ -4,7 +4,7 @@ const logo_url = 'https://d1gkd65id2ekzc.cloudfront.net/test.jpg';
 const Logo = (boxProps: BoxProps) => {
   return (
     <Box {...boxProps}>
-        <Image src={logo_url} alt="logo" />
+        <Image maxH="50px" borderRadius='md' src={logo_url} alt="logo" />
     </Box>
   )
 }
