+++ website/src/components/NavBar.tsx
@@ -4,9 +4,11 @@ import {
     Tab,
     Box,
     Divider,
+    Text
 } from "@chakra-ui/react"
 import React, { FunctionComponent } from "react";
 import { Link } from "react-router-dom";
+import Logo from "./Logo";
 
 type Props = {
     selectedIndex: 0 | 1 | 2
@@ -19,7 +21,9 @@ const NavBar: FunctionComponent<Props> = ({selectedIndex}) => {
                 my="2" 
                 variant='soft-rounded' 
                 index={selectedIndex}>
-                <TabList justifyContent='center'>
+                <TabList justifyContent='center' alignItems='center' ml='-125'>
+                    <Text>(Logo)</Text>
+                    <Logo mx="3" />    
                     <Tab>
                         <Link to="/">Home</Link>
                     </Tab>
